Adjusted Dockerfile and configure.js to use external directories if necessary

+++ bin/configure.js
@@ -2,7 +2,7 @@
 
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2019 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -30,23 +30,26 @@ const [defaultGitConfigurationUrl, defaultGitConfigurationVersion] = pkg.depende
 );
 const gitConfigurationUrl = process.env.WIRE_CONFIGURATION_REPOSITORY || defaultGitConfigurationUrl;
 const gitConfigurationVersion = process.env.WIRE_CONFIGURATION_REPOSITORY_VERSION || defaultGitConfigurationVersion;
+const configDirName = process.env.WIRE_CONFIGURATION_EXTERNAL_DIR || 'config';
 
-console.log(
-  `Loading configuration version "${gitConfigurationVersion}" for project "${pkg.name}" from "${gitConfigurationUrl}"`
-);
-
-const configDirName = 'config';
 const configDir = resolve(configDirName);
 const src = resolve(configDir, pkg.name, 'content');
 const projectDir = resolve('./');
 const dest = resolve(projectDir, 'resource');
 const ignoreList = ['.DS_Store'];
 
-console.log(`Cleaning config directory "${configDir}"`);
-fs.removeSync(configDir);
-execSync(`git clone --single-branch -b ${gitConfigurationVersion} ${gitConfigurationUrl} ${configDirName}`, {
-  stdio: [0, 1],
-});
+if (!process.env.WIRE_CONFIGURATION_EXTERNAL_DIR) {
+  console.log(
+    `Loading configuration version "${gitConfigurationVersion}" for project "${
+      pkg.name
+    }" from "${gitConfigurationUrl}" \
+     and cleaning config directory "${configDir}"`,
+  );
+  fs.removeSync(configDir);
+  execSync(`git clone -b ${gitConfigurationVersion} ${gitConfigurationUrl} ${configDirName}`, {stdio: [0, 1]});
+} else {
+  console.log(`Using external config directory "${configDir}"`);
+}
 
 // Copy .env file configuration
 fs.copySync(resolve(configDir, pkg.name, '.env'), resolve(projectDir, '.env'));
